# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Мониторинг интернета ВКО",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/devices/enrollment-codes": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Выдать одноразовый код установки агента для школы",
        "description": "Неизвестный school_id — 422.",
        "operationId": "create_enrollment_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentCodeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentCodeIssued"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/devices/{device_id}": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Карточка устройства",
        "operationId": "get_device",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          },
          "404": {
            "description": "Устройство не найдено",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{device_id}/measurements": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "История замеров устройства, новые сверху",
        "operationId": "list_device_measurements",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          },
          {
            "name": "period_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Начало периода по measured_at, включительно",
              "title": "Period From"
            },
            "description": "Начало периода по measured_at, включительно"
          },
          {
            "name": "period_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Конец периода по measured_at, не включается",
              "title": "Period To"
            },
            "description": "Конец периода по measured_at, не включается"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы, с 1",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы, с 1"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Размер страницы"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementListItemPage"
                }
              }
            }
          },
          "404": {
            "description": "Устройство не найдено",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{device_id}/block": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Заблокировать устройство: запросы агента отклоняются, история остаётся",
        "operationId": "block_device",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          },
          "404": {
            "description": "Устройство не найдено",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{device_id}/unblock": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Разблокировать устройство",
        "operationId": "unblock_device",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          },
          "404": {
            "description": "Устройство не найдено",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации запроса",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "online",
          "offline"
        ]
      },
...
      "DeviceDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Device ID"
          },
          "device_uid": {
            "type": "string",
            "title": "Device Uid",
            "description": "Постоянный идентификатор ПК от агента"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "monitoring_point_id": {
            "type": "integer",
            "title": "Monitoring Point Id"
          },
          "monitoring_point_name": {
            "type": "string",
            "title": "Monitoring Point Name"
          },
          "room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room",
            "description": "Кабинет точки мониторинга"
          },
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "line_status": {
            "$ref": "#/components/schemas/LineStatus"
          },
          "agent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At",
            "description": "Последняя связь с агентом"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          },
          "current_status": {
            "$ref": "#/components/schemas/SchoolStatus",
            "description": "По правилу T-16: heartbeat в рабочие часы и последний замер; вне рабочих часов — no_data"
          },
          "latest_measurement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestMeasurement"
              },
              {
                "type": "null"
              }
            ]
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          },
          "school_id": {
            "type": "integer",
            "title": "School Id"
          },
          "school_code": {
            "type": "string",
            "title": "School Code",
            "description": "School ID"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_uid",
          "hostname",
          "monitoring_point_id",
          "monitoring_point_name",
          "room",
          "line_id",
          "line_status",
          "agent_version",
          "last_seen_at",
          "status",
          "current_status",
          "latest_measurement",
          "os",
          "school_id",
          "school_code",
          "school_name",
          "registered_at"
        ],
        "title": "DeviceDetail",
        "description": "Device card (ТЗ п. 4): the list row plus the school, the OS and the registration time."
      },
...
      "DeviceStatus": {
        "type": "string",
        "enum": [
          "active",
          "blocked"
        ]
      },
      "EnrollmentCodeCreate": {
        "properties": {
          "school_id": {
            "type": "integer",
            "title": "School Id"
          }
        },
        "type": "object",
        "required": [
          "school_id"
        ],
        "title": "EnrollmentCodeCreate",
        "description": "Request for a one-time agent installation code for a school (plan.md §4.1)."
      },
      "EnrollmentCodeIssued": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Код для параметра установки ENROLL_CODE; повторно не показывается",
            "examples": [
              "VKO-7F3K-92QD"
            ]
          },
          "school_id": {
            "type": "integer",
            "title": "School Id"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At",
            "description": "Окончание срока действия; по умолчанию 7 дней с выдачи (ADR-005)",
            "examples": [
              "2026-09-24T04:00:00Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "school_id",
          "expires_at"
        ],
        "title": "EnrollmentCodeIssued",
        "description": "Issued installation code: shown once, the server keeps only its hash (ADR-005)."
      },
...
      "IfaceType": {
        "type": "string",
        "enum": [
          "ethernet",
          "wifi",
          "other"
        ]
      },
      "LatestMeasurement": {
        "properties": {
          "measured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Measured At",
            "description": "Момент замера на ПК"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "download_mbps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Mbps"
          },
          "upload_mbps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Mbps"
          },
          "ping_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ping Ms"
          },
          "jitter_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jitter Ms"
          },
          "packet_loss_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Packet Loss Pct"
          },
          "iface_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IfaceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "thresholds_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThresholdValues"
              },
              {
                "type": "null"
              }
            ],
            "description": "Пороги, применённые при оценке этого замера (ТЗ п. 11)"
          },
          "quality_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Пусто, если замер не оценён"
          }
        },
        "type": "object",
        "required": [
          "measured_at",
          "connection_status",
          "download_mbps",
          "upload_mbps",
          "ping_ms",
          "jitter_ms",
          "packet_loss_pct",
          "iface_type",
          "thresholds_snapshot",
          "quality_status"
        ],
        "title": "LatestMeasurement",
        "description": "Values of the latest measurement shown in cards and lists (ТЗ п. 4, п. 13)."
      },
...
      "LineStatus": {
        "type": "string",
        "enum": [
          "main",
          "reserve",
          "disabled"
        ]
      },
...
      "MeasurementListItem": {
        "properties": {
          "measured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Measured At",
            "description": "Момент замера на ПК"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "download_mbps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Mbps"
          },
          "upload_mbps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Mbps"
          },
          "ping_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ping Ms"
          },
          "jitter_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jitter Ms"
          },
          "packet_loss_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Packet Loss Pct"
          },
          "iface_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IfaceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "thresholds_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThresholdValues"
              },
              {
                "type": "null"
              }
            ],
            "description": "Пороги, применённые при оценке этого замера (ТЗ п. 11)"
          },
          "quality_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Пусто, если замер не оценён"
          },
          "measurement_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Measurement Uuid"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At",
            "description": "Приём сервером; разница с measured_at — время в офлайн-очереди"
          },
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "external_ip": {
            "anyOf": [
              {
                "type": "string",
                "format": "ipvanyaddress"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ip"
          },
          "server": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server",
            "description": "Сервер и метод замера"
          },
          "agent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          },
          "contract_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Ok",
            "description": "Факт не ниже договорной скорости линии; пусто — договорных значений нет"
          }
        },
        "type": "object",
        "required": [
          "measured_at",
          "connection_status",
          "download_mbps",
          "upload_mbps",
          "ping_ms",
          "jitter_ms",
          "packet_loss_pct",
          "iface_type",
          "thresholds_snapshot",
          "quality_status",
          "measurement_uuid",
          "received_at",
          "line_id",
          "duration_s",
          "external_ip",
          "server",
          "agent_version",
          "contract_ok"
        ],
        "title": "MeasurementListItem",
        "description": "Measurement in the history: raw values, the line and the server evaluation (plan.md §5).\n\n``line_id`` is the line the measurement went through, possibly the reserve one (plan.md §4.5)."
      },
      "MeasurementListItemPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeasurementListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "MeasurementListItemPage",
        "description": "Page of the measurement history of a device, newest first."
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный машинный код ошибки",
            "examples": [
              "not_found"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Название ошибки для алерта",
            "examples": [
              "Не найдено"
            ]
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "examples": [
              404
            ]
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Текст для человека"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "Путь запроса"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "Error of any endpoint; clients branch on ``type``, ``detail`` is for a human."
      },
      "QualityStatus": {
        "type": "string",
        "enum": [
          "normal",
          "unstable",
          "critical",
          "offline"
        ]
      },
...
      "SchoolStatus": {
        "type": "string",
        "enum": [
          "normal",
          "unstable",
          "critical",
          "offline",
          "no_data"
        ]
      },
...
      "ThresholdValues": {
        "properties": {
          "download_min_mbps": {
            "type": "number",
            "minimum": 0.0,
            "title": "Download Min Mbps",
            "examples": [
              20
            ]
          },
          "upload_min_mbps": {
            "type": "number",
            "minimum": 0.0,
            "title": "Upload Min Mbps",
            "examples": [
              20
            ]
          },
          "ping_max_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Ping Max Ms",
            "examples": [
              100
            ]
          },
          "jitter_max_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Jitter Max Ms",
            "examples": [
              30
            ]
          },
          "packet_loss_max_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Packet Loss Max Pct",
            "examples": [
              2
            ]
          }
        },
        "type": "object",
        "required": [
          "download_min_mbps",
          "upload_min_mbps",
          "ping_max_ms",
          "jitter_max_ms",
          "packet_loss_max_pct"
        ],
        "title": "ThresholdValues",
        "description": "Limits of the «Норма» status; base values of ТЗ п. 11 are the examples."
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "examples": [
              "items[3].ping_ms"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Ожидается число"
            ]
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "description": "One invalid field: dotted path from the request root, e.g. ``items[3].ping_ms``."
      },
      "ValidationProblem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Стабильный машинный код ошибки",
            "examples": [
              "validation_error"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Название ошибки для алерта",
            "examples": [
              "Ошибка валидации"
            ]
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "examples": [
              422
            ]
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Текст для человека"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "Путь запроса"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "errors"
        ],
        "title": "ValidationProblem",
        "description": "Request validation error (HTTP 422, ``type = validation_error``)."
      }
    },
    "securitySchemes": {
      "DeviceToken": {
        "type": "apiKey",
        "description": "Токен устройства: `Authorization: Device <token>` (ADR-005)",
        "in": "header",
        "name": "Authorization"
      },
      "RefreshCookie": {
        "type": "apiKey",
        "description": "Refresh-токен панели в httpOnly cookie",
        "in": "cookie",
        "name": "refresh_token"
      },
      "BearerAuth": {
        "type": "http",
        "description": "Access-токен панели: `Authorization: Bearer <jwt>`, живёт 15 минут",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}